README.fedora: Clarify dnf v.s. yum plugin
Mention that we pull in a plugin for dnf, not just yum.

Also, neither yum nor dnf plugin require special handling when removing
etckeeper.  The dnf plugin completely ignores errors in the post handler.
The yum plugin post checks that /usr/bin/etckeeper has not been removed
before it attempts to run it.